Rename OperationPipe to DataSourceOperationsQueue Simplify interface of DataSourceOperationsQueue
(cherry picked from commit 92a5133df38a146766f2917d643e98b375e5b57f)
